% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmd3.0_studies}
\alias{cmd3.0_studies}
\title{a snapshot of studies that are being curated}
\format{
data.frame
}
\usage{
cmd3.0_studies
}
\description{
a snapshot of studies that are being curated
}
\examples{
names(cmd3.0_studies)
head(cmd3.0_studies, 3)
}
\keyword{datasets}
