% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampatts.R
\name{samp_atts_by_study}
\alias{samp_atts_by_study}
\title{obtain sample attributes for a study.}
\usage{
samp_atts_by_study(srp = "SRP082656", size = 500, cursor = NULL, ...)
}
\arguments{
\item{srp}{character(1) SRP tag for a study in NCBI SRA}

\item{size}{numeric(1) maximum number of records to return}

\item{cursor}{character(1) the value of the cursor returned in a previous call; see note.}

\item{\dots}{not currently used}
}
\description{
obtain sample attributes for a study.
}
\note{
If the number of records available in query response exceeds the value of \code{size}, the
returned tibble will have an attribute 'cursor' which can be retrieved and used in
subsequent calls to this function.

If \code{size} is zero, a list is returned with information on the number of records available.
}
\examples{
t1 = samp_atts_by_study(size=10)
t1
attr(t1, "cursor")
samp_atts_by_study(size=0)
}
